fix: resolve static analysis findings and improve test coverage
- remote-control: replace C-style casts with reinterpret_cast (MISRA)
- remote-control: fix ssize_t -> uint16_t narrowing conversion in remote.h
- remote-control: add explicit static_cast for htons() port argument
- CanSpecConfigLoader tests: add 7 edge-case tests (empty value, invalid
  number, hex 0X prefix, whitespace trimming, empty list, fallback paths)
- ConfigParser tests: add 4 branch coverage tests (multi-entry, float
  precision, bool edge cases, delimiter-only lines)

Coverage improvement (common module):
  Line:   92.5% -> 99.6%
  Branch: 60.7% -> 69.3%

Release build: 0 warnings across all modules (-Werror clean)

+++ katech-automotive/projects/autonomous-driving-stack/modules/common/include/modules/remote-control/remote.h
@@ -26,7 +26,7 @@ class CANClass
         struct canfd_frame FrameFd;
     private:
         int sock;
-        uint16_t nbytes;
+        ssize_t nbytes;
         struct sockaddr_can addr;
         struct ifreq ifr;
         socklen_t addrlen;
@@ -44,7 +44,7 @@ public:
 private:
     int sock;
     struct sockaddr_in Addr;
-    uint16_t nbytes;
+    ssize_t nbytes;
     socklen_t addrlen;
 };
